refactor: rename event package to messaging and update references
